renamed files to match the classes that they contain
